Clients: Rename client/commands to cli #7253
