+++ lib/rucio/cli/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from rucio.client.commands import command  # noqa: F401
+from rucio.cli import command  # noqa: F401
