+++ lib/rucio/cli/bin_legacy/__init__.py
