+++ lib/rucio/cli/upload.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 import click
 
-from rucio.client.commands.bin_legacy.rucio import upload
-from rucio.client.commands.utils import Arguments
+from rucio.cli.bin_legacy.rucio import upload
+from rucio.cli.utils import Arguments
 from rucio.common.config import config_get_float
